Add usage of new auth header for apps (#7920)
