+++ saleor/core/auth_backend.py
@@ -1,7 +1,8 @@
 from django.contrib.auth.backends import ModelBackend
 
 from ..account.models import User
-from .jwt import get_token_from_request, get_user_from_access_token
+from .auth import get_token_from_request
+from .jwt import get_user_from_access_token
 
 
 class JSONWebTokenBackend(ModelBackend):
